Utilities refactoring
